refactor: Simplify dependency injection implementation
