--- ai_chat_app/lib/app/di/data/annotations.dart
@@ -1,6 +1,0 @@
-import 'package:injectable/injectable.dart';
-
-const development = Environment('development');
-const staging = Environment('staging');
-const production = Environment('production');
-const testing = Environment('testing');
