--- ai_chat_app/lib/app/di/data/config/test_config.config.dart
@@ -1,27 +1,0 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-// **************************************************************************
-// InjectableConfigGenerator
-// **************************************************************************
-
-// ignore_for_file: type=lint
-// coverage:ignore-file
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:get_it/get_it.dart' as _i174;
-import 'package:injectable/injectable.dart' as _i526;
-
-extension GetItInjectableX on _i174.GetIt {
-// initializes the registration of main-scope dependencies inside of GetIt
-  _i174.GetIt initTest({
-    String? environment,
-    _i526.EnvironmentFilter? environmentFilter,
-  }) {
-    _i526.GetItHelper(
-      this,
-      environment,
-      environmentFilter,
-    );
-    return this;
-  }
-}
